api/requirements/base.txt
# Core FastAPI e dipendenze essenziali
annotated-types==0.7.0
anyio==4.10.0
bcrypt==5.0.0
cffi==2.0.0
click==8.2.1
cryptography==46.0.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.2
more-itertools==10.8.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing_extensions==4.15.0
uvicorn==0.35.0

# Database - includi entrambi per flessibilità
psycopg2-binary==2.9.10